Change Pub/Sub exercise order
