+++ courses/data_analysis/deepdive/pubsub-exercises/exercise1/labs/run.sh
@@ -10,7 +10,5 @@ gcloud pubsub subscriptions delete $OLD_SUBSCRIPTION
 echo "Create new subscription."
 gcloud pubsub subscriptions create --topic pubsub-e2e-example $SUBSCRIPTION
 echo $SUBSCRIPTION > .last_subscription
-echo "Starting publisher and publishing 1M messages."
-java -Xmx1024m -cp target/pubsub.jar com.google.cloud.sme.pubsub.Publisher -p $PROJECT
-echo "Starting subscriber."
-java -Xmx1024m -cp target/pubsub.jar com.google.cloud.sme.pubsub.Subscriber -p $PROJECT --subscription $SUBSCRIPTION
+echo "Starting publisher and publishing 10M messages."
+java -Xmx1024m -cp target/pubsub.jar com.google.cloud.sme.pubsub.Publisher $@ -p $PROJECT
