+++ courses/data_analysis/deepdive/pubsub-exercises/exercise1/labs/src/main/java/com/google/cloud/sme/pubsub/Publisher.java
@@ -20,11 +20,12 @@ import com.beust.jcommander.Parameter;
 import com.google.api.core.ApiFuture;
 import com.google.cloud.sme.Entities;
 import com.google.cloud.sme.common.ActionReader;
-import com.google.cloud.sme.common.ActionUtils;
 import com.google.cloud.sme.common.FileActionReader;
 import com.google.protobuf.ByteString;
 import com.google.pubsub.v1.ProjectTopicName;
 import com.google.pubsub.v1.PubsubMessage;
+import java.nio.ByteBuffer;
+import java.util.Map;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.atomic.AtomicLong;
@@ -38,39 +39,30 @@ public class Publisher {
         required = true,
         description = "The Google Cloud Pub/Sub project in which the topic exists.")
     public String project = null;
-
-    @Parameter(
-        names = {"--ordered", "-o"},
-        required = false,
-        description = "Whether or not to publish messages with an ordering key.")
-    public Boolean ordered = false;
   }
 
   private static final String SOURCE_DATA = "actions.csv";
-  private static final String SEQUENCE_NUM_KEY = "sequence_num";
+  private static final String TIMESTAMP_KEY = "publish_time";
   private static final String TOPIC = "pubsub-e2e-example";
-  private static final int MESSAGE_COUNT = 1000;
+  private static final int MESSAGE_COUNT = 10000000;
 
   private final Args args;
   private com.google.cloud.pubsub.v1.Publisher publisher;
   private ActionReader actionReader;
   private AtomicLong awaitedFutures;
   private ExecutorService executor = Executors.newCachedThreadPool();
-  private ByteString extraInfo;
+
+  private Stats stats;
 
   private Publisher(Args args, ActionReader actionReader) {
     this.args = args;
     this.actionReader = actionReader;
     this.awaitedFutures = new AtomicLong();
-    byte[] extraBytes = new byte[1024];
-    this.extraInfo = ByteString.copyFrom(extraBytes);
+    this.stats = new Stats();
 
     ProjectTopicName topic = ProjectTopicName.of(args.project, TOPIC);
     com.google.cloud.pubsub.v1.Publisher.Builder builder =
         com.google.cloud.pubsub.v1.Publisher.newBuilder(topic);
-    if (args.ordered) {
-      builder.setEnableMessageOrdering(true);
-    }
     try {
       this.publisher = builder.build();
     } catch (Exception e) {
@@ -79,21 +71,21 @@ public class Publisher {
     }
   }
 
-  private void Publish(long sequenceNum, Entities.Action publishAction) {
+  private void Publish(Entities.Action publishAction) {
     awaitedFutures.incrementAndGet();
-    publishAction = Entities.Action.newBuilder(publishAction).setExtraInfo(this.extraInfo).build();
     final long publishTime = DateTime.now().getMillis();
-    PubsubMessage.Builder messageBuilder =
+    ByteBuffer buffer = ByteBuffer.allocate(Long.BYTES);
+    buffer.putLong(publishAction.getItemId());
+    PubsubMessage message =
         PubsubMessage.newBuilder()
-            .setData(ActionUtils.encodeAction(publishAction))
-            .putAttributes(SEQUENCE_NUM_KEY, Long.toString(sequenceNum));
-    if (args.ordered) {
-      messageBuilder.setOrderingKey(Long.toString(publishAction.getUserId()));
-    }
-    ApiFuture<String> response = publisher.publish(messageBuilder.build());
+            .setData(ByteString.copyFrom(buffer))
+            .putAttributes(TIMESTAMP_KEY, Long.toString(publishTime))
+            .build();
+    ApiFuture<String> response = publisher.publish(message);
     response.addListener(
         () -> {
           try {
+            stats.recordLatency(DateTime.now().getMillis() - publishTime);
             response.get();
           } catch (Exception e) {
             System.out.println("Could not publish a message: " + e);
@@ -105,15 +97,17 @@ public class Publisher {
   }
 
   private void run() {
+    stats.start();
     awaitedFutures.incrementAndGet();
 
+    long count = 0;
     Entities.Action nextAction = actionReader.next();
     for (int i = 0; i < MESSAGE_COUNT; ++i) {
-      Publish(i, nextAction);
-      nextAction = actionReader.next();
+      Publish(nextAction);
       if ((i + 1) % 100000 == 0) {
         System.out.println("Published " + (i + 1) + " messages.");
       }
+      nextAction = actionReader.next();
     }
     awaitedFutures.decrementAndGet();
   }
@@ -126,6 +120,7 @@ public class Publisher {
     } catch (InterruptedException e) {
       System.out.println("Error while waiting for completion: " + e);
     }
+    stats.stop(DateTime.now().getMillis());
     executor.shutdownNow();
   }
 
@@ -135,6 +130,18 @@ public class Publisher {
     } catch (Exception e) {
       System.out.println("Error while shutting down: " + e);
     }
+  }
+
+  private void printStats() {
+    Map<Double, Long> latencies = stats.getLatencies();
+    Long throughput = stats.getThroughputMessages();
+    System.out.println("Publishing complete.");
+    System.out.println("Published " + stats.getMessageCount() + " messages.");
+    System.out.println("Publish latency");
+    for (Map.Entry<Double, Long> latency : latencies.entrySet()) {
+      System.out.println(" " + latency.getKey() + "th percentile: " + latency.getValue() + "ms");
+    }
+    System.out.println("Throughput: " + throughput + " messages/s");
   }
 
   public static void main(String[] args) {
@@ -148,6 +155,7 @@ public class Publisher {
     p.run();
     p.waitForPublishes();
     p.shutdown();
+    p.printStats();
     System.exit(0);
   }
 }
