+++ courses/data_analysis/deepdive/pubsub-exercises/exercise1/solution/build.sh
