+++ courses/data_analysis/deepdive/pubsub-exercises/exercise1/solution/src/main/java/com/google/cloud/sme/common/FileItemReader.java
