+++ courses/data_analysis/deepdive/pubsub-exercises/exercise1/solution/src/main/proto/entities.proto
