+++ courses/data_analysis/deepdive/pubsub-exercises/exercise2/labs/pom.xml
@@ -2,7 +2,7 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.google.cloud.sme</groupId>
-  <artifactId>pubsub-sme-exercise-2-lab</artifactId>
+  <artifactId>pubsub-sme-exercise-3-lab</artifactId>
   <packaging>jar</packaging>
   <version>1.0-SNAPSHOT</version>
   <name>example</name>
