+++ courses/data_analysis/deepdive/pubsub-exercises/exercise2/labs/src/main/java/com/google/cloud/sme/pubsub/Publisher.java
@@ -20,12 +20,11 @@ import com.beust.jcommander.Parameter;
 import com.google.api.core.ApiFuture;
 import com.google.cloud.sme.Entities;
 import com.google.cloud.sme.common.ActionReader;
+import com.google.cloud.sme.common.ActionUtils;
 import com.google.cloud.sme.common.FileActionReader;
 import com.google.protobuf.ByteString;
 import com.google.pubsub.v1.ProjectTopicName;
 import com.google.pubsub.v1.PubsubMessage;
-import java.nio.ByteBuffer;
-import java.util.Map;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.atomic.AtomicLong;
@@ -44,21 +43,21 @@ public class Publisher {
   private static final String SOURCE_DATA = "actions.csv";
   private static final String TIMESTAMP_KEY = "publish_time";
   private static final String TOPIC = "pubsub-e2e-example";
-  private static final int MESSAGE_COUNT = 10000000;
+  private static final int MESSAGE_COUNT = 1000;
 
   private final Args args;
   private com.google.cloud.pubsub.v1.Publisher publisher;
   private ActionReader actionReader;
   private AtomicLong awaitedFutures;
   private ExecutorService executor = Executors.newCachedThreadPool();
-
-  private Stats stats;
+  private ByteString extraInfo;
 
   private Publisher(Args args, ActionReader actionReader) {
     this.args = args;
     this.actionReader = actionReader;
     this.awaitedFutures = new AtomicLong();
-    this.stats = new Stats();
+    byte[] extraBytes = new byte[102400];
+    this.extraInfo = ByteString.copyFrom(extraBytes);
 
     ProjectTopicName topic = ProjectTopicName.of(args.project, TOPIC);
     com.google.cloud.pubsub.v1.Publisher.Builder builder =
@@ -73,19 +72,17 @@ public class Publisher {
 
   private void Publish(Entities.Action publishAction) {
     awaitedFutures.incrementAndGet();
+    publishAction = Entities.Action.newBuilder(publishAction).setExtraInfo(this.extraInfo).build();
     final long publishTime = DateTime.now().getMillis();
-    ByteBuffer buffer = ByteBuffer.allocate(Long.BYTES);
-    buffer.putLong(publishAction.getItemId());
     PubsubMessage message =
         PubsubMessage.newBuilder()
-            .setData(ByteString.copyFrom(buffer))
+            .setData(ActionUtils.encodeAction(publishAction))
             .putAttributes(TIMESTAMP_KEY, Long.toString(publishTime))
             .build();
     ApiFuture<String> response = publisher.publish(message);
     response.addListener(
         () -> {
           try {
-            stats.recordLatency(DateTime.now().getMillis() - publishTime);
             response.get();
           } catch (Exception e) {
             System.out.println("Could not publish a message: " + e);
@@ -97,16 +94,12 @@ public class Publisher {
   }
 
   private void run() {
-    stats.start();
     awaitedFutures.incrementAndGet();
 
     long count = 0;
     Entities.Action nextAction = actionReader.next();
     for (int i = 0; i < MESSAGE_COUNT; ++i) {
       Publish(nextAction);
-      if ((i + 1) % 100000 == 0) {
-        System.out.println("Published " + (i + 1) + " messages.");
-      }
       nextAction = actionReader.next();
     }
     awaitedFutures.decrementAndGet();
@@ -120,7 +113,6 @@ public class Publisher {
     } catch (InterruptedException e) {
       System.out.println("Error while waiting for completion: " + e);
     }
-    stats.stop(DateTime.now().getMillis());
     executor.shutdownNow();
   }
 
@@ -130,18 +122,6 @@ public class Publisher {
     } catch (Exception e) {
       System.out.println("Error while shutting down: " + e);
     }
-  }
-
-  private void printStats() {
-    Map<Double, Long> latencies = stats.getLatencies();
-    Long throughput = stats.getThroughputMessages();
-    System.out.println("Publishing complete.");
-    System.out.println("Published " + stats.getMessageCount() + " messages.");
-    System.out.println("Publish latency");
-    for (Map.Entry<Double, Long> latency : latencies.entrySet()) {
-      System.out.println(" " + latency.getKey() + "th percentile: " + latency.getValue() + "ms");
-    }
-    System.out.println("Throughput: " + throughput + " messages/s");
   }
 
   public static void main(String[] args) {
@@ -155,7 +135,6 @@ public class Publisher {
     p.run();
     p.waitForPublishes();
     p.shutdown();
-    p.printStats();
     System.exit(0);
   }
 }
