+++ courses/data_analysis/deepdive/pubsub-exercises/exercise3/labs/run_ordered.sh
